Working on the routine
Should be about done with the initial schedule and run of the simulator.
Currently working on the adjustment of the schedule due to lead times of
phantoms from the first.  This part should end up recurring until no
lead time issues remain.  But it looks like it will have a risk of
looping forever since there is nothing preventing it from moving dates
back forward.