Overhauling structure
Trying a different structure.  Instead of scheduling each production
area based on MO priority separately, then checking for problems and
trying again, the scheduler will just set one order in place at a time.
It starts with a priority of SOs and MOs so sales are the first priority
driver.  Then it tries to schedule one order each loop.
This will likely have a significantly longer runtime than the previous
structure.  But it will be far more accurate to priority requirements
and log better parent structure for the orders.